python/langgraph0dot6dot5__cpython3dot9dot6/pyproject.toml
[project]
name = "langgraph0dot6dot5-cpython3dot9dot6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.6"
dependencies = [
    "datadog-api-client==2.25.0",
    "fastapi==0.116.1",
    "httpx==0.28.1",
    "langchain-core>=0.3.67",
    "langchain-openai==0.2.11",
    "langgraph==0.6.5",
    "pytest==7.4.4",
    "python-dotenv==1.0.1",
    "tavily-python==0.3.3",
    "uvicorn==0.24.0",
]
